fix: enhance cached options
