+++ src/vite/type.d.ts
@@ -1,7 +1,7 @@
 declare namespace NodeJS {
   interface Process {
     electronApp: import('node:child_process').ChildProcess
-    cachedOptions: import('./types').ResolvedOptions
+    cachedElectronOptions?: import('vite-plugin-electron/multi-env').MultiEnvElectronOptions[]
     __electron_path: string | undefined
     __bytecode_compiler_path: string | undefined
   }
