+++ src/vite/types.ts
@@ -319,44 +319,3 @@ export interface BuildVersionOptions {
   generateSignature: NonNullable<GeneratorOverrideFunctions['generateSignature']>
   generateUpdateJson: NonNullable<GeneratorOverrideFunctions['generateUpdateJson']>
 }
-
-export interface UpdaterParsedOptions {
-  buildAsarOption: BuildAsarOptions
-  buildVersionOption: BuildVersionOptions
-  entryOutDir: string
-}
-
-export interface ResolvedOptions {
-  /** Whether is in build mode */
-  isBuild: boolean
-  /**
-   * Project root directory, resolved from vite config or falls back to process.cwd()
-   */
-  root: string
-  /** Entry build options */
-  entry: ElectronWithUpdaterOptions['entry']
-  /** Main process build options */
-  main: ElectronWithUpdaterOptions['main']
-  /** Preload build options */
-  preload: ElectronWithUpdaterOptions['preload']
-  /** Whether to generate sourcemap */
-  sourcemap: boolean
-  /** Whether to minify the code */
-  minify: boolean
-  /** Whether to generate version json */
-  buildVersionJson: boolean | undefined
-  /** Whether to use notBundle plugin */
-  notBundleOption: boolean | NotBundleOptions
-  isESM: boolean
-  finalExternal: (string | RegExp)[]
-  bytecodeOptions: BytecodeOptions | undefined
-  buildAsarOption: BuildAsarOptions
-  buildVersionOption: BuildVersionOptions
-  entryOutDir: string
-  outputNames: {
-    entryFileNames: string
-    chunkFileNames: string
-    assetFileNames: string
-  }
-  define: Record<string, string>
-}
